requirements.txt
aiogram==2.25.1
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
cbor2==5.4.6
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorzero==1.1
cryptography==39.0.1
distro==1.5.0
Flask==2.2.2
frozenlist==1.3.3
gpiozero==1.6.2
idna==2.6
importlib-metadata==5.2.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
luma.core==2.4.0
luma.led-matrix==1.7.0
magic-filter==1.0.9
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.1
opencv-contrib-python==4.5.3.56
opencv-python==4.6.0.66
picamera==1.13
picamera2==0.3.9
pidng==4.0.9
piexif==1.1.3
Pillow==9.5.0
pycparser==2.21
pyftdi==0.54.0
PyMLX90614==0.0.4
pypng==0.20220715.0
pyserial==3.5
python-apt==2.2.1
python-dotenv==0.21.0
python-prctl==1.7
pytz==2023.3
pyusb==1.2.1
pyzbar==0.1.9
qrcode==7.4.2
requests==2.18.4
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
simplejpeg==1.6.4
six==1.16.0
smbus2==0.4.2
spidev==3.5
sqlparse==0.4.3
ssh-import-id==5.10
telepot==12.7
toml==0.10.1
typing-extensions==4.5.0
urllib3==1.22
v4l2-python3==0.3.2
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
